Add project description and clean up code
SaikaVA is a website with the goal to create little arcade games with the dragonmaid theme. The games will be played in the browser and
the player should be able to fill in his username, when he tops one of the highscores. The top 10 highscores will be saved in a databank 
displayed on the website. The arcade games should be simple and may be reinvented in an mobile app later on. Change the dragonmaid 
characters, images and sounds to prevent copyright issues.