Update azure-pipelines.yml for Azure Pipelines
By default, PHP_CodeSniffer will exit with a non-zero code if any errors or warnings are found. If you want to display errors to the user, but still return with a zero exit code if no warnings are found, you can set the ignore_errors_on_exit configuration option. This option is typically used by automated build tools so that a list of errors can be generated without failing the build.